+++ e2e/runner/setup_evm.go
@@ -213,3 +213,29 @@ func (r *E2ERunner) DeployZetaConnectorNative(ensureTxReceipt func(tx *ethtypes.
 		txZetaConnectorNativeHash.Hash().Hex(),
 	)
 }
+
+// DeployTestDAppV2EVM deploys the test DApp V2 contract for EVM
+func (r *E2ERunner) DeployTestDAppV2EVM() {
+	ensureTxReceipt := func(tx *ethtypes.Transaction, failMessage string) {
+		receipt := utils.MustWaitForTxReceipt(r.Ctx, r.EVMClient, tx, r.Logger, r.ReceiptTimeout)
+		r.requireTxSuccessful(receipt, failMessage)
+	}
+
+	testDAppV2Addr, txTestDAppV2, _, err := testdappv2.DeployTestDAppV2(
+		r.EVMAuth,
+		r.EVMClient,
+		false,
+		r.GatewayEVMAddr,
+	)
+	require.NoError(r, err)
+	ensureTxReceipt(txTestDAppV2, "TestDAppV2 deployment failed")
+
+	// Initialize the test dapp contract instance
+	r.TestDAppV2EVMAddr = testDAppV2Addr
+	r.TestDAppV2EVM, err = testdappv2.NewTestDAppV2(testDAppV2Addr, r.EVMClient)
+	require.NoError(r, err)
+
+	isZetaChain, err := r.TestDAppV2EVM.IsZetaChain(&bind.CallOpts{})
+	require.NoError(r, err)
+	require.False(r, isZetaChain)
+}